Udate - smaller changes to fishtank and footer. Moved store folder to src. Edited READMEs
