+++ frontend/src/data/my-attendance.json
@@ -2,7 +2,7 @@
   "hero": {
     "h1": "My attendance",
     "pre": "Your private space",
-    "intro": "Forgot what you choice for food or who to bring as your plus one? We've got you.",
+    "intro": "Forgot what you chose for food or who to bring as your plus one? We've got you.",
     "paragraph": "This is your page. Here you'll find your RSVP and maybe other info that's only for you. Let us know if anything is missing. Haven't RSVP'd yet? Please, do so.",
     "image": "/cody-black-VDb0wxbfG6k-unsplash_1920x1263.jpg",
     "cta": {
