+++ frontend/src/sections/FishtankSection.jsx
@@ -13,7 +13,7 @@ export const FishtankSection = () => {
               speed: -15,
               children: (
                 <div className="absolute inset-0 flex items-center justify-center ">
-                  <p className="mx-8 text-center font-cormorant text-2xl font-light leading-10 text-white">
+                  <p className=" mx-8 max-w-screen-md text-center font-cormorant text-2xl font-light leading-10 text-white lg:text-3xl">
                     “Once in a while, right in the middle of an ordinary life,
                     love gives us a fairy tale.”
                   </p>
